Fixed issues with incorrect refactoring of helper_functions
/utils back again, and previous utils.py -> helper_functions.py moved to proper place
and refactored paths to helper_functions.
